ai_pipeline/requirements.txt
tensorflow>=2.20.0
pandas>=2.3.1
numpy>=2.3.2
pyserial>=3.5
